feat: :boom: Replace `[VAR]_PATH` with `[VAR]_FILE`
According to Docker, there is a convention to use a `_FILE` suffix
for variables pointing to a file path containing the value [1].

Since we're especially supporting the file route because of these kinds
of features [2], we should follow this convention.

[1]: https://docs.docker.com/compose/how-tos/use-secrets/#advanced
[2]: https://docs.docker.com/engine/swarm/secrets/#build-support-for-docker-secrets-into-your-images

BREAKING CHANGE: The `[VAR]_PATH` environment variable has been replaced with `[VAR]_FILE`.